refactor: split cron delivery planning from sending
